Fix typo, needed to return v in validators for descriptions to fix NoneType errors.
