+++ tests/data/basic_descriptor.json
@@ -1,5 +1,5 @@
 {
     "name": "descriptor",
-    "descriptions": [],
-    "next_fib": 1
+    "next_fib": 1,
+    "descriptions": []
 }
